feat: usa MDAWidget
