+++ src/pymmcore_widgets/_mda/_positions_table_widget.py
@@ -163,6 +163,7 @@ class PositionTable(QGroupBox):
         lbl = QLabel("Z Stage:")
         lbl.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
         self.z_stage_combo = QComboBox()
+        self.z_stage_combo.setSizeAdjustPolicy(QComboBox.AdjustToContents)
         self.z_stage_combo.currentTextChanged.connect(self._on_z_stage_combo_changed)
         cbox_wdg_layout.addWidget(lbl)
         cbox_wdg_layout.addWidget(self.z_stage_combo)
